+++ Tests/StoragePerfTest.cpp
@@ -36,7 +36,7 @@ void run_ddl(const string& input_str) {
   list<Parser::Stmt*> parse_trees;
   string last_parsed;
   CHECK_EQ(parser.parse(input_str, parse_trees, last_parsed), 0);
-  CHECK_EQ(parse_trees.size(), 1);
+  CHECK_EQ(parse_trees.size(), size_t(1));
   auto stmt = parse_trees.front();
   unique_ptr<Stmt> stmt_ptr(stmt);  // make sure it's deleted
   Parser::DDLStmt* ddl = dynamic_cast<Parser::DDLStmt*>(stmt);
